pyproject.toml
[project]
name = "cinereserve-api"
version = "0.1.0"
description = ""
authors = [
    {name = "thiagoeduardoc",email = "thiagoeduardoc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "psycopg2-binary (>=2.9.9)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]

[tool.poetry]
packages = [{include = "cinereserve_api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
